Fix memeverseLauncher abi

+++ src/router/interfaces/IMemeverseLauncher.sol
@@ -5,5 +5,5 @@ pragma solidity ^0.8.28;
  * @title MemeverseLauncher interface
  */
 interface IMemeverseLauncher {
-    function genesis(uint256 verseId, uint256 amountInUPT, address user) external;
+    function genesis(uint256 verseId, uint128 amountInUPT, address user) external;
 }
